pyproject.toml
[project]
name = "voice-driven-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
    "pyaudio>=0.2.14",
    "pystray>=0.19.5",
    "python-dotenv>=1.1.1",
    "pyttsx3>=2.99",
    "requests>=2.32.4",
    "speechrecognition>=3.14.3",
]
